Rename canvas -> content-views
