+++ src/content-views/__tests__/colors-test.js
