+++ src/content-views/canvasUtils.js
