+++ src/content-views/colors.js
