+++ src/content-views/views/FlamechartView.js
