+++ src/content-views/views/ReactEventsView.js
