+++ src/content-views/views/ReactMeasuresView.js
